Open file and then write to it.
Trying to support older versions of puppet and ruby.  Apparently the static
method File.write didn't exist in the version of ruby we're using, so we
have to do it this way instead.  Also took the opportunity to set
restrictive permissions on the file when it's created, since node_encrypt
depends on a subsequent file resource to adjust things like permissions.